requirements.txt
Django~=5.2
asgiref~=3.8.1
sqlparse~=0.5.3
django-debug-toolbar~=5.1.0
python-dotenv~=1.1.0
psycopg2-binary
django-environ
GDAL==3.6.2
Pillow
PyJWT==2.10.1
sorl-thumbnail==12.11.0
openrouteservice==2.3.3
